Moving storage implementations into separate packages
Fixes https://github.com/kubernetes-incubator/service-catalog/issues/199
