+++ pkg/controller/storage/mem/storage_broker.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package storage
+package mem
 
 import (
 	"errors"
