+++ pkg/controller/storage/tpr/storage_service_class.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package storage
+package tpr
 
 import (
 	"log"
